fix searchBar in homepage ui
